Refactor progress tests into separate file
